medical_chatbot/requirements.txt
Django==4.2
# OpenAI
openai==2.8.1
python-dotenv==1.0.0